alterações propostas no review do Yuri e junção das estruturas de armazenamento na classe Database
